refactorying -> moved Cell hadlers to cells file
