+++ libraries/cells/gridAllergeni.py
@@ -1,4 +1,4 @@
-from libraries.gridCell import Cell
+from libraries.cells.gridCell import Cell
 import tkinter as tk
 import tkinter.font as TkFont
